refactor: Remove abstract llm_d toolbox commands
